customers/serializers.py
from rest_framework import serializers
from .models import Customer

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ['id', 'name', 'phone', 'email', 'created_at', 'updated_at', 'is_active']
        read_only_fields = ['created_at', 'updated_at', 'is_active']
