requirements.txt
django>=4.2
djangorestframework
drf-yasg
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
